Fixed connecting curves offset issue
The problem was that the node components were being resized to 0x0 when 
their tabs were not in focus so the offsets to output components were 
off and the entire thing was incorrect. Now, if the resize width or 
previous width is zero (meaning it is from one of these tab switching 
resizes) it will not change the curves.